% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChildren.R
\name{addChildren}
\alias{addChildren}
\title{Add children blocks to parent blocks that have not yet been posted.}
\usage{
addChildren(parentBlock, ...)
}
\arguments{
\item{parentBlock}{Data frame. The parent block to add children to.}

\item{...}{Data frame representing Notion content to add to parent block. May accept multiple blocks.}
}
\value{
Data frame object representing the Notion block object.
}
\description{
\code{addChildren} facilitates adding children blocks to parent blocks that
have not yet been posted (providing the block type allows children).
Currently this includes "paragraph", "bulleted_list_item", "numbered_list_item",
"toggle", "to_do", "quote", "callout", "synced_block", "template", "column",
"child_page", and "child_database".
}
\details{
To add children to blocks that are already posted, use \code{patchAppendBlocks}.
}
\examples{
toggleBlock <- createTextBlock("toggle","Outer toggle text")
stuffForToggle <- createTextBlock("paragraph", "inner toggle text")
moreStuffForToggle <- createTextBlock("quote", "Woof woof")
finalToggleBlock <- addChildren(toggleBlock, stuffForToggle, moreStuffForToggle)
## Pipe-friendly
finalToggleBlock <- createTextBlock("toggle","Outer toggle text") \%>\%
addChildren(createTextBlock("paragraph", "inner toggle text"), moreStuffForToggle)
}
