% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReferenceBlock.R
\name{createReferenceBlock}
\alias{createReferenceBlock}
\title{Create a new url-dependent block to send to Notion.}
\usage{
createReferenceBlock(type, url)
}
\arguments{
\item{type}{String. Currently supports the following types: image, video, file, pdf, bookmark.}

\item{url}{String. The reference URL where the content is stored.}
}
\value{
Data frame object representing the Notion block object.
}
\description{
\code{createReferenceBlock} creates objects for blocks that reference a url. See the type parameter for a full list.
Captions are not currently supported.
}
\examples{
url = "https://developers.notion.com"
bookmarkBlock <- createReferenceBlock("bookmark", url)

}
