% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTextBlock.R
\name{createTextBlock}
\alias{createTextBlock}
\title{Create a new block to send to Notion.}
\usage{
createTextBlock(type, text)
}
\arguments{
\item{type}{String. Currently supports the following types: paragraph, heading_1, heading_2, heading_3,
callout, quote, bulleted_list_item, numbered_list_item, to_do, toggle.}

\item{text}{String. The text content of the block.}
}
\value{
Data frame object representing the Notion block object.
}
\description{
\code{createTextBlock} creates objects for most text-focused block types. See the "type" parameter for a full list.
}
\examples{
createTextBlock("paragraph","This is the text of a new paragraph block")
createTextBlock("quote","Do not be miserable before it is time. - Seneca")
}
