% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_handlers.R
\name{get_notion_key}
\alias{get_notion_key}
\title{Get and return NOTION_KEY}
\usage{
get_notion_key()
}
\value{
NOTION_KEY as a string
}
\description{
Returns the NOTION_KEY environment variable. If no key is set, the function will return a warning.
}
\examples{
\dontrun{get_notion_key()}
}
