% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchAppendBlocks.R
\name{patchAppendBlocks}
\alias{patchAppendBlocks}
\title{Append blocks to a published block or a page.}
\usage{
patchAppendBlocks(content_id, ...)
}
\arguments{
\item{content_id}{Character string. The Notion page id or URL for the parent block to append to.}

\item{...}{Data frames representing Notion blocks. May enter more than one.}
}
\value{
httr2 response
}
\description{
Append block objects to published blocks/pages. Pass blocks to append as individual unnamed arguments.
This function will format and send the PATCH request to Notion.
}
\examples{
testBlock1 <- createTextBlock("paragraph", "My first test block")
testBlock2 <- createTextBlock("quote", "Another test block")
\dontrun{patchAppendBlocks(page_id, testBlock1, testBlock2)}


}
