% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postPage.R
\name{postPage}
\alias{postPage}
\title{Post a page block to Notion.}
\usage{
postPage(page, page_id = NULL)
}
\arguments{
\item{page}{List object representing a page block in Notion. Use \code{createPage} to create a new page block.}

\item{page_id}{Character string. The Notion page id or URL for the parent page of the page being posted.}
}
\value{
httr2 response
}
\description{
Post a page block to Notion.
}
\examples{
aNewPage <- createPage(title = "My brand new page")
page_id = "31d8db33bxyxy8d09xyxyb42758886ef"
\dontrun{postPage(aNewPage, page_id)}
}
