% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_handlers.R
\name{set_notion_key}
\alias{set_notion_key}
\title{Set NOTION_KEY as an environment variable}
\usage{
set_notion_key(key = NULL)
}
\arguments{
\item{key}{A string.}
}
\value{

}
\description{
Call \code{set_notion_key()} to save your Notion secret key as an environment variable. Leave the \code{key} parameter empty to call \code{askpass} and enter your secret there.
}
\examples{
\dontrun{set_notion_key()}
}
