% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChildren.R
\name{addChildren}
\alias{addChildren}
\title{Add children blocks to parent blocks.}
\usage{
addChildren(parentBlock, children)
}
\arguments{
\item{parentBlock}{Data frame. The parent block to add children to.}

\item{children}{Data frame. The child block(s) you are adding.}
}
\value{
Data frame object representing the Notion block object.
}
\description{
\code{addChildren} facilitates adding children blocks to parent blocks
whose type allows it. Currently this includes "paragraph",
"bulleted_list_item", "numbered_list_item", "toggle", "to_do", "quote",
"callout", "synced_block", "template", "column", "child_page", and
"child_database".
}
\examples{
toggleBlock <- createTextBlock("toggle","Outer toggle text")
stuffForToggle <- createTextBlock("paragraph", "inner toggle text")
toggleBlock <- addChildren(toggleBlock, stuffForToggle)
## Pipe-friendly
toggleBlock <- createTextBlock("toggle","Outer toggle text") \%>\%
addChildren(createTextBlock("paragraph", "inner toggle text"))
}
