% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPage.R
\name{createPage}
\alias{createPage}
\title{Create a page in Notion.}
\usage{
createPage(page_id = NULL, title, content = NULL, postPage = FALSE)
}
\arguments{
\item{page_id}{Character string. The Notion page id of the parent page for your new page.}

\item{title}{Character string. The title of your new page.}

\item{content}{Data frame. The content block to be added to your new page.}

\item{postPage}{Boolean. If set to TRUE, the final page block will be posted to Notion.}
}
\value{
List object representing a Notion page block.
}
\description{
This function facilitates basic page creation in Notion. Create a block using createTextBlock or a
similar function and pass to the content parameter to create a page with content. Returns a list object
representing a Notion page block.
}
\examples{
newParagraphBlock <- createTextBlock("paragraph", "This is a new paragraph block")
page_id = "31d8db33bxyxy8d09xyxyb42758886ef"
\dontrun{createPage(page_id, "My new page title", newParagraphBlock)}
}
